+++ test/runtests.jl
@@ -6,15 +6,8 @@ import InteractiveUtils: versioninfo
 import LinearAlgebra: BLAS
 
 # there seem to be processor-specific issues and knowing this is helpful
-println(versioninfo())
-@static if VERSION ≥ v"1.7.0-DEV.620"
-    println(BLAS.get_config())
-else
-    @show BLAS.vendor()
-    if startswith(string(BLAS.vendor()), "openblas")
-        println(BLAS.openblas_get_config())
-    end
-end
+@info sprint(versioninfo)
+@info BLAS.get_config()
 
 @testset "Aqua" begin
     # we can't check for unbound type parameters
@@ -29,13 +22,13 @@ include("UniformBlockDiagonal.jl")
 include("linalg.jl")
 include("matrixterm.jl")
 include("FactorReTerm.jl")
+include("grouping.jl")
 include("pls.jl")
 include("pirls.jl")
 include("gausshermite.jl")
 include("fit.jl")
 include("missing.jl")
 include("likelihoodratiotest.jl")
-include("grouping.jl")
 include("bootstrap.jl")
 include("mime.jl")
 include("optsummary.jl")